forgot to merge in this bugfix

+++ cascadenik-compile.py
@@ -44,7 +44,8 @@ def main(src_file, dest_file, **kwargs):
         f.close()
         
     # manually unlinking seems to be required on windows
-    os.unlink(dest_file)
+    if os.path.exists(dest_file):
+        os.unlink(dest_file)
     os.rename(tmp_file, dest_file)
     return 0
